+++ CHANGELOG.md
@@ -5,7 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Changed configuration description (EnableSpawnProtection).
+- Added "Force Despawn Bosses" to configuration, so that bosses spawned via effects can be forcibly despawned if all
+  players are dead.
+- Changed configuration description (EnableSpawnProtection) so that it is easier to understand.
 - Increased "Explode Inventory" drop chance value from 10 to 15.
 
 ## [1.0.0] - 2022-07-23